// apps/web/src/lib/api-client/models/managedUserDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * AppStart API
 * API REST do AppStart - Especificação OpenAPI 3.0
 * OpenAPI spec version: 1.0.0
 */
import type { ManagedUserDtoRole } from './managedUserDtoRole';

export interface ManagedUserDto {
  id: string;
  name: string;
  email: string;
  role: ManagedUserDtoRole;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}